addressing review comments - testing remaining
